# docs/openapi.json
{"openapi": "3.1.0", "info": {"title": "Nifty 100 Analytics API", "version": "1.0.0"}, "paths": {"/api/v1/health": {"get": {"summary": "Get Health", "operationId": "get_health_api_v1_health_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/api/v1/companies": {"get": {"summary": "Get Companies", "operationId": "get_companies_api_v1_companies_get", "parameters": [{"name": "sector", "in": "query", "required": false, "schema": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Sector"}}, {"name": "market_cap_category", "in": "query", "required": false, "schema": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Market Cap Category"}}, {"name": "search", "in": "query", "required": false, "schema": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Search"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/companies/{ticker}": {"get": {"summary": "Get Company Profile", "operationId": "get_company_profile_api_v1_companies__ticker__get", "parameters": [{"name": "ticker", "in": "path", "required": true, "schema": {"type": "string", "title": "Ticker"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/companies/{ticker}/pl": {"get": {"summary": "Get Company Pl", "operationId": "get_company_pl_api_v1_companies__ticker__pl_get", "parameters": [{"name": "ticker", "in": "path", "required": true, "schema": {"type": "string", "title": "Ticker"}}, {"name": "from_year", "in": "query", "required": false, "schema": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "From Year"}}, {"name": "to_year", "in": "query", "required": false, "schema": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "To Year"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/companies/{ticker}/bs": {"get": {"summary": "Get Company Bs", "operationId": "get_company_bs_api_v1_companies__ticker__bs_get", "parameters": [{"name": "ticker", "in": "path", "required": true, "schema": {"type": "string", "title": "Ticker"}}, {"name": "from_year", "in": "query", "required": false, "schema": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "From Year"}}, {"name": "to_year", "in": "query", "required": false, "schema": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "To Year"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/companies/{ticker}/cashflow": {"get": {"summary": "Get Company Cf", "operationId": "get_company_cf_api_v1_companies__ticker__cashflow_get", "parameters": [{"name": "ticker", "in": "path", "required": true, "schema": {"type": "string", "title": "Ticker"}}, {"name": "from_year", "in": "query", "required": false, "schema": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "From Year"}}, {"name": "to_year", "in": "query", "required": false, "schema": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "To Year"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/companies/{ticker}/ratios": {"get": {"summary": "Get Company Ratios", "operationId": "get_company_ratios_api_v1_companies__ticker__ratios_get", "parameters": [{"name": "ticker", "in": "path", "required": true, "schema": {"type": "string", "title": "Ticker"}}, {"name": "year", "in": "query", "required": false, "schema": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Year"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/companies/{ticker}/tearsheet": {"get": {"summary": "Get Company Tearsheet", "operationId": "get_company_tearsheet_api_v1_companies__ticker__tearsheet_get", "parameters": [{"name": "ticker", "in": "path", "required": true, "schema": {"type": "string", "title": "Ticker"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/screener": {"get": {"summary": "Run Screener", "operationId": "run_screener_api_v1_screener_get", "parameters": [{"name": "min_roe", "in": "query", "required": false, "schema": {"anyOf": [{"type": "number"}, {"type": "null"}], "title": "Min Roe"}}, {"name": "max_de", "in": "query", "required": false, "schema": {"anyOf": [{"type": "number"}, {"type": "null"}], "title": "Max De"}}, {"name": "min_fcf", "in": "query", "required": false, "schema": {"anyOf": [{"type": "number"}, {"type": "null"}], "title": "Min Fcf"}}, {"name": "sector", "in": "query", "required": false, "schema": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Sector"}}, {"name": "min_rev_cagr_5yr", "in": "query", "required": false, "schema": {"anyOf": [{"type": "number"}, {"type": "null"}], "title": "Min Rev Cagr 5Yr"}}, {"name": "min_pat_cagr_5yr", "in": "query", "required": false, "schema": {"anyOf": [{"type": "number"}, {"type": "null"}], "title": "Min Pat Cagr 5Yr"}}, {"name": "max_pe", "in": "query", "required": false, "schema": {"anyOf": [{"type": "number"}, {"type": "null"}], "title": "Max Pe"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/sectors": {"get": {"summary": "Get Sectors", "operationId": "get_sectors_api_v1_sectors_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/api/v1/sectors/{sector}/companies": {"get": {"summary": "Get Sector Companies", "operationId": "get_sector_companies_api_v1_sectors__sector__companies_get", "parameters": [{"name": "sector", "in": "path", "required": true, "schema": {"type": "string", "title": "Sector"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/peers/{group_name}": {"get": {"summary": "Get Peer Group", "operationId": "get_peer_group_api_v1_peers__group_name__get", "parameters": [{"name": "group_name", "in": "path", "required": true, "schema": {"type": "string", "title": "Group Name"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/companies/{ticker}/peers/compare": {"get": {"summary": "Get Company Peer Compare", "operationId": "get_company_peer_compare_api_v1_companies__ticker__peers_compare_get", "parameters": [{"name": "ticker", "in": "path", "required": true, "schema": {"type": "string", "title": "Ticker"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/market-cap/{ticker}": {"get": {"summary": "Get Historical Valuation", "operationId": "get_historical_valuation_api_v1_market_cap__ticker__get", "parameters": [{"name": "ticker", "in": "path", "required": true, "schema": {"type": "string", "title": "Ticker"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/portfolio/stats": {"get": {"summary": "Get Portfolio Stats", "operationId": "get_portfolio_stats_api_v1_portfolio_stats_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/api/v1/companies/{ticker}/documents": {"get": {"summary": "Get Company Documents", "operationId": "get_company_documents_api_v1_companies__ticker__documents_get", "parameters": [{"name": "ticker", "in": "path", "required": true, "schema": {"type": "string", "title": "Ticker"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}}, "components": {"schemas": {"HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}, "input": {"title": "Input"}, "ctx": {"type": "object", "title": "Context"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}}}